requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.36
google-auth==2.33.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
